% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporter-Inference.R
\name{docinfeR}
\alias{docinfeR}
\title{Inference Report in .docx file}
\usage{
docinfeR(path = "", var.type = "all")
}
\arguments{
\item{path}{(Optional) A character vector with the path to data file. If empty character string (""), interface will appear to choose file.}

\item{var.type}{(Optional) The type of variables to perform analysis, with possible values: "all", "numeric", "integer", "double", "factor", "character".}
}
\value{
The output
  will be a document in the temp folder (tempdir()).
}
\description{
This R Package asks for a .csv file with data and returns a report (.docx) with Inference Report concerning all possible combinations of variables (i.e. columns).
}
\examples{
\donttest{
library(docinfeR)
data(iris)
dir = tempdir()
write.csv(iris,file=paste(dir,"iriscsvfile.csv",sep=""))
docinfeR(path=paste(dir,"iriscsvfile.csv",sep=""))
}
}
\references{
\insertAllCited{}
}
